utils: create tester script
Users can now use:
- `python -m saturn_engine.utils.tester`

This script allows for easily testing saturn objects using
tests defined in YAML files.

For now, we only support the SaturnPipelineTest kind but we can add
other test types in the future.

The `simple.yaml` config was also adapted to add an example of an API key
resource.

You can use `./example/test_pipeline` to see it in action.